handle customize namespace for the neoload serivce remove the keptn.neoload.engine.yaml to implement the workload.yaml structure.
Have the ability to let neoload web use existing zone id

Add the suppor for the dynatrace integration

+++ src/main/java/com/keptn/neotys/testexecutor/messageHandler/NeoLoadHandler.java
@@ -2,6 +2,7 @@ package com.keptn.neotys.testexecutor.messageHandler;
 
 import com.keptn.neotys.testexecutor.EventSender.NeoLoadEndEvent;
 import com.keptn.neotys.testexecutor.KeptnEvents.KeptnEventFinished;
+import com.keptn.neotys.testexecutor.NeoLoadFolder.ProjectSettings;
 import com.keptn.neotys.testexecutor.NeoLoadFolder.datamodel.*;
 import com.keptn.neotys.testexecutor.NeoLoadFolder.variables.NeoLoadModel;
 import com.keptn.neotys.testexecutor.NeoLoadFolder.variables.NlConstants;
@@ -86,14 +87,14 @@ public class NeoLoadHandler {
                     logger.debug("getNeoLoadTest - no able to deserialize the yaml file");
                     listFuture.fail(new NeoLoadSerialException("Unable to deserialize YAML file "));
                 }
-                if (neoLoadDataModel.getSteps().size() < 0) {
+                if (neoLoadDataModel.getWorkloads().size() < 0) {
                     logger.debug("getNeoLoadTest - there is no testing steps");
                     listFuture.fail(new NeoLoadJgitExeption("There is no testing steps define "));
 
                 }
 
                 final ArrayList<NeoLoadTestStep> neoLoadTestSteps = new ArrayList<>();
-                listFuture.complete(neoLoadDataModel.getSteps());
+                listFuture.complete(neoLoadDataModel.getWorkloads());
             } else {
                 logger.error("No Ressrouce " + NEOLOAD_CONFIG_FILE + " found for project " + keptnEventFinished.getProject() + " and stage " + keptnEventFinished.getStage());
                 listFuture.fail(new NeoLoadJgitExeption("No Ressrouce " + NEOLOAD_CONFIG_FILE + " found for project " + keptnEventFinished.getProject() + " and stage " + keptnEventFinished.getStage()));
@@ -150,7 +151,7 @@ public class NeoLoadHandler {
         return projectwithoutnlp.stream().map(project->{return new File(project).getName();}).collect(Collectors.joining(","));
     }
 
-    private String createZipFile(List<String> projectPath, String projectName, Optional<List<Constants>> constant_variables) throws IOException, NeoLoadJgitExeption {
+    private String createZipFile(List<String> projectPath, String projectName, Optional<List<Constants>> constant_variables,Optional<String> dynatraceTenant,Optional<String> dynatraceToken) throws IOException, NeoLoadJgitExeption {
         Path path = Paths.get(gitfolder.toAbsolutePath().toString()+TMP_NEOLOAD_FOLDER);
         if(!Files.exists(path))
             Files.createDirectory(path);
@@ -182,7 +183,17 @@ public class NeoLoadHandler {
             NeoLoadModel model=new NeoLoadModel(constant_variables.get().stream().map(constants -> {
                 return new NlConstants(constants);
             }).collect(Collectors.toList()));
-            Yaml yaml = new Yaml();
+            Yaml yaml;
+            if(dynatraceTenant.isPresent()&& dynatraceToken.isPresent())
+            {
+                model.setProjectSettings(new ProjectSettings( true,dynatraceTenant.get(),dynatraceToken.get()));
+                yaml=new Yaml(model.getProjectSettings().getRepresenter());
+            }
+            else
+            {
+               yaml=new Yaml();
+            }
+
 
             tempfile=Optional.of(path.toAbsolutePath().toString()+"/"+keptnEventFinished.getService()+"."+keptncontext+YAML_EXTENSION);
             try {
@@ -195,6 +206,26 @@ public class NeoLoadHandler {
                 error.add(e);
             }
         }
+        else
+        {
+            if(dynatraceTenant.isPresent()&& dynatraceToken.isPresent())
+            {
+                NeoLoadModel model=new NeoLoadModel();
+                model.setProjectSettings(new ProjectSettings( true,dynatraceTenant.get(),dynatraceToken.get()));
+                Yaml yaml=new Yaml(model.getProjectSettings().getRepresenter());
+
+                tempfile=Optional.of(path.toAbsolutePath().toString()+"/"+keptnEventFinished.getService()+"."+keptncontext+YAML_EXTENSION);
+                try {
+                    yaml.dump(model, new FileWriter(tempfile.get()));
+                    logger.debug("Constant yaml file created : " + tempfile.get());
+                }
+                catch (IOException e)
+                {
+                    logger.error("Issue to create the constant yaml file",e);
+                    error.add(e);
+                }
+            }
+        }
          if(error.size()>0)
         {
             throw new NeoLoadJgitExeption("several tecnical error : "+ error.stream().map(e -> {
@@ -205,7 +236,7 @@ public class NeoLoadHandler {
         return compressNLProject(path.toAbsolutePath().toString(),projectName+keptnEventFinished.getService());
 
     }
-    private NeoLoadWebTest RunTest(File zipfile, NeoLoadTest test, Optional<String> nlapi, Optional<String> nlapitoken, Optional<String> nlurl, Optional<String> uploadurl, Optional<String> nlzoneid, int size) throws ApiException, NeoLoadJgitExeption {
+    private NeoLoadWebTest RunTest(File zipfile, NeoLoadTestStep test, Optional<String> nlapi, Optional<String> nlapitoken, Optional<String> nlurl, Optional<String> uploadurl, Optional<String> nlzoneid, int size) throws ApiException, NeoLoadJgitExeption {
        if(!nlapi.isPresent())
            throw new NeoLoadJgitExeption("No API URL Defined. installtion of the neoload service has not been configured properly");
 
@@ -246,8 +277,8 @@ public class NeoLoadHandler {
             nlWebApiClient.setApiKey(nlapitoken.get());
             runtimeApi=new RuntimeApi(nlWebApiClient);
 
-            RunTestDefinition runTestDefinition = runtimeApi.getTestsRun(KEPTN_EVENT_URL+"_"+keptnEventFinished.getProject()+"_"+keptnEventFinished.getService()+"_"+test.getScenario(), projectDefinition.getProjectId(), test.getScenario(), test.getDescription(),getAsCodeFiles(test.getProject()),null,null,null,null,nlzoneid.get(),nlzoneid.get()+":"+String.valueOf(size));
-            NeoLoadWebTest neoLoadWebTest=new NeoLoadWebTest(runTestDefinition.getTestId(),NLWEB_PROTOCOL+nlurl.get() + "/#!trend/?scenario=" + test.getScenario() + "&limit=-1&project=" + projectDefinition.getProjectId(),NLWEB_PROTOCOL+nlurl.get() + "/#!result/" + runTestDefinition.getTestId() + "/overview");
+            RunTestDefinition runTestDefinition = runtimeApi.getTestsRun(KEPTN_EVENT_URL+"_"+keptnEventFinished.getProject()+"_"+keptnEventFinished.getService()+"_"+test.getProperties().getScenario(), projectDefinition.getProjectId(), test.getProperties().getScenario(), test.getDescription(),getAsCodeFiles(test.getScript().getProject()),null,null,null,null,nlzoneid.get(),nlzoneid.get()+":"+String.valueOf(size));
+            NeoLoadWebTest neoLoadWebTest=new NeoLoadWebTest(runTestDefinition.getTestId(),NLWEB_PROTOCOL+nlurl.get() + "/#!trend/?scenario=" + test.getProperties().getScenario() + "&limit=-1&project=" + projectDefinition.getProjectId(),NLWEB_PROTOCOL+nlurl.get() + "/#!result/" + runTestDefinition.getTestId() + "/overview");
 
 
             logger.info("Trending URL : " + neoLoadWebTest.getTrendingurl());
@@ -310,57 +341,73 @@ public class NeoLoadHandler {
         }
         return directoryToBeDeleted.delete();
     }
-    private void runNLScenario(NeoLoadTest test, Vertx rxvertx, CloudEvent<Object> receivedEvent)
+    private void runNLScenario(NeoLoadTestStep test, Vertx rxvertx, CloudEvent<Object> receivedEvent)
     {
         final NeoLoadKubernetesClient neoLoadKubernetesClient=new NeoLoadKubernetesClient(keptncontext);
-        List<String> machinelist = null;
-        try {
-            if (test.checkProject(gitfolder)) {
-                machinelist=test.getNameOfLGtoStart(gitfolder);
-                logger.debug("This scenario "+test.getScenario()+" requires " + machinelist.size()+ " NeoLoad LoadGenerator");
-                //----start machines----
+        int numberofInstances = 0;
+        String zoneid = null;
+        if(test.getInfrastructure().isManagedbyKeptn())
+        {
+            List<String> machinelist = null;
+            try {
+                if (test.checkProject(gitfolder)) {
+                    machinelist = test.getNameOfLGtoStart();
+                    logger.debug("This scenario " + test.getProperties().getScenario() + " requires " + machinelist.size() + " NeoLoad LoadGenerator");
+                    //----start machines----
 
-                neoLoadKubernetesClient.deployController();
+                    neoLoadKubernetesClient.deployController();
 
-                List<String> error=new ArrayList<>();
-                machinelist.stream().forEach(machine->{
-                    try {
-                        neoLoadKubernetesClient.deployLG(machine);
-                    } catch (NeoLoadJgitExeption neoLoadJgitExeption) {
-                        logger.error("deployment issue ",neoLoadJgitExeption);
-                        error.add(neoLoadJgitExeption.getMessage());
+                    List<String> error = new ArrayList<>();
+                    machinelist.stream().forEach(machine -> {
+                        try {
+                            neoLoadKubernetesClient.deployLG(machine);
+                        } catch (NeoLoadJgitExeption neoLoadJgitExeption) {
+                            logger.error("deployment issue ", neoLoadJgitExeption);
+                            error.add(neoLoadJgitExeption.getMessage());
 
-                    }
-                });
+                        }
+                    });
+                    if(error.size()>0)
+                        throw new NeoLoadJgitExeption("Issue while deploying LGs");
 
-                if(error.size()>0)
-                    throw new NeoLoadJgitExeption("Issue while deploying LGs");
+                    zoneid=neoLoadKubernetesClient.getNeoloadZoneid().get();
+                    numberofInstances=machinelist.size();
+                }
+                }
+            catch (NeoLoadJgitExeption e)
+            {
+                logger.error("runNLScenario exepption ",e);
+            }
+            catch (NeoLoadSerialException e)
+            {
+                logger.error("runNLScenario exepption ",e);
+            } catch (IOException e) {
+                logger.error("runNLScenario exepption ",e);
+            }
+        }
+        else {
+            numberofInstances=test.getInfrastructure().getNumberOfMachine().intValue();
+            zoneid=test.getInfrastructure().getZoneId();
 
-                List<String> projectspath=test.getProject().stream().map(project -> project.getPath()).collect(Collectors.toList());
-                String zipfilepath=createZipFile(projectspath, keptnEventFinished.getProject(),Optional.ofNullable(test.getConstant_variables()));
+        }
+        try
+        {
+                List<String> projectspath=test.getScript().getProject().stream().map(project -> project.getPath()).collect(Collectors.toList());
+                String zipfilepath=createZipFile(projectspath, keptnEventFinished.getProject(),Optional.ofNullable(test.getProperties().getConstant_variables()),neoLoadKubernetesClient.getDynatrace_tenant(),neoLoadKubernetesClient.getDynatrace_api_token());
 
                 Thread.sleep(20000);
 
-                NeoLoadWebTest loadWebTest=RunTest(new File(zipfilepath),test,neoLoadKubernetesClient.getNeoloadweb_apiurl(),Optional.ofNullable(neoLoadKubernetesClient.getNeoloadAPitoken()),neoLoadKubernetesClient.getNeoloadweb_url(),neoLoadKubernetesClient.getNeoloadweb_uploadurl(),neoLoadKubernetesClient.getNeoloadZoneid(),machinelist.size());
+                NeoLoadWebTest loadWebTest=RunTest(new File(zipfilepath),test,neoLoadKubernetesClient.getNeoloadweb_apiurl(),Optional.ofNullable(neoLoadKubernetesClient.getNeoloadAPitoken()),neoLoadKubernetesClient.getNeoloadweb_url(),neoLoadKubernetesClient.getNeoloadweb_uploadurl(),Optional.ofNullable(zoneid),numberofInstances);
                 keptnEventFinished.setTestid(loadWebTest.getTestid());
                 keptnEventFinished.setNeoloadURL(loadWebTest.getTesturl());
                 ///---
                 NeoLoadEndEvent endEvent=new NeoLoadEndEvent(logger,eventid,rxvertx);
-                endEvent.endevent(keptnEventFinished,extensions,receivedEvent);
+                endEvent.endevent(keptnEventFinished,extensions,receivedEvent,neoLoadKubernetesClient.getKeptn_NAMESPACE());
                 //--send end event-------------
 
-            }
+
         }
-        catch (NeoLoadJgitExeption e)
-        {
-            logger.error("runNLScenario exepption ",e);
-        }
-        catch (NeoLoadSerialException e)
-        {
-            logger.error("runNLScenario exepption ",e);
-        } catch (IOException e) {
-            logger.error("runNLScenario exepption ",e);
-        } catch (ApiException | InterruptedException e) {
+        catch (ApiException | InterruptedException e) {
             logger.error("RUnNLScenario , api exception",e);
         }catch (Exception e)
         {
@@ -372,12 +419,16 @@ public class NeoLoadHandler {
             {
                 try {
                     //delete infra
-                    if (machinelist != null) {
-                        machinelist.stream().forEach(machine -> {
-                            neoLoadKubernetesClient.deleteLG(machine);
-                        });
+                    if (numberofInstances >0) {
+                        if (test.getInfrastructure().isManagedbyKeptn()) {
+                            test.getNameOfLGtoStart().stream().forEach(machine-> {
+
+                                    neoLoadKubernetesClient.deleteLG(machine);
+
+                              });
+                        }
+                        neoLoadKubernetesClient.deleteController();
                     }
-                    neoLoadKubernetesClient.deleteController();
                 }
                 catch (Exception e)
                 {
@@ -403,19 +454,19 @@ public class NeoLoadHandler {
                 StringBuilder error=new StringBuilder();
 
                 //---for each test start test -----
-                neoLoadTestStepList.stream().filter(neoLoadTestStep -> neoLoadTestStep.getStep().getTeststrategy().equalsIgnoreCase(keptnEventFinished.getTeststrategy())).forEach(step->{
-                    logger.debug("Clonning repo :"+step.getStep().getRepository());
+                neoLoadTestStepList.stream().filter(neoLoadTestStep -> neoLoadTestStep.getTeststrategy().equalsIgnoreCase(keptnEventFinished.getTeststrategy())).forEach(step->{
+                    logger.debug("Clonning repo :"+step.getScript().getRepository());
 
                     try {
-                        gitfolder=getNeoLoadTestFolder(getGitHubFolder(step.getStep().getRepository()),step.getStep().getBranch());
+                        gitfolder=getNeoLoadTestFolder(getGitHubFolder(step.getScript().getRepository()),step.getScript().getBranch());
                     } catch (IOException e) {
                         error.append("Technical Error while retrieveing repository "+e.getMessage());
                     } catch (NeoLoadJgitExeption neoLoadJgitExeption) {
                         error.append("Technical Error while retrieveing repository "+neoLoadJgitExeption.getMessage());
                     }
-                    logger.debug("Running step : "+step.getStep().getScenario());
-                    runNLScenario(step.getStep(),rxvertx,receivedEvent);
-                    logger.debug("end step : "+step.getStep().getScenario());
+                    logger.debug("Running step : "+step.getProperties().getScenario());
+                    runNLScenario(step,rxvertx,receivedEvent);
+                    logger.debug("end step : "+step.getProperties().getScenario());
                     try {
                         deleteGitFolder();
                     } catch (IOException e) {
